refactor: 컨트롤러 테스트 중복 코드 분리
