+++ anonymous-board/src/test/java/com/example/anonymousboard/post/controller/PostControllerTest.java
@@ -1,14 +1,13 @@
 package com.example.anonymousboard.post.controller;
 
-import static com.example.anonymousboard.util.ApiDocumentUtils.getDocumentRequest;
-import static com.example.anonymousboard.util.ApiDocumentUtils.getDocumentResponse;
-import static com.example.anonymousboard.util.DocumentFormatGenerator.getConstraints;
+import static com.example.anonymousboard.util.apidocs.ApiDocumentUtils.getDocumentRequest;
+import static com.example.anonymousboard.util.apidocs.ApiDocumentUtils.getDocumentResponse;
+import static com.example.anonymousboard.util.apidocs.DocumentFormatGenerator.getConstraints;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doThrow;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
@@ -31,47 +30,18 @@ import com.example.anonymousboard.post.exception.InvalidPostKeywordException;
 import com.example.anonymousboard.post.exception.InvalidTitleException;
 import com.example.anonymousboard.post.exception.PostErrorCode;
 import com.example.anonymousboard.post.exception.PostNotFoundException;
-import com.example.anonymousboard.post.service.PostService;
-import com.example.anonymousboard.support.AuthInterceptor;
-import com.example.anonymousboard.support.token.JwtTokenProvider;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.example.anonymousboard.util.controller.ControllerTest;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
 
-@WebMvcTest(PostController.class)
-@ExtendWith(RestDocumentationExtension.class)
-public class PostControllerTest {
-
-    @Autowired
-    MockMvc mockMvc;
-
-    @Autowired
-    ObjectMapper objectMapper;
-
-    @MockBean
-    PostService postService;
-
-    @MockBean
-    AuthInterceptor authInterceptor;
-
-    @MockBean
-    JwtTokenProvider jwtTokenProvider;
+public class PostControllerTest extends ControllerTest {
 
     PagePostsResponse pagePostsResponse;
 
@@ -88,13 +58,7 @@ public class PostControllerTest {
     PostResponse keywordPostResponse2;
 
     @BeforeEach
-    void setUp(WebApplicationContext webApplicationContext, RestDocumentationContextProvider restDocumentation) {
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-                .apply(documentationConfiguration(restDocumentation))
-                .build();
-        given(authInterceptor.preHandle(any(), any(), any()))
-                .willReturn(true);
-
+    void setUp() {
         postResponse1 = PostResponse.builder()
                 .id(1L)
                 .title("제목1")
