--- anonymous-board/src/test/java/com/example/anonymousboard/util/ApiDocumentUtils.java
@@ -1,26 +1,0 @@
-package com.example.anonymousboard.util;
-
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.modifyUris;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-
-import org.springframework.restdocs.operation.preprocess.OperationRequestPreprocessor;
-import org.springframework.restdocs.operation.preprocess.OperationResponsePreprocessor;
-
-public class ApiDocumentUtils {
-
-    public static OperationRequestPreprocessor getDocumentRequest() {
-        return preprocessRequest(
-                modifyUris()
-                        .scheme("https")
-                        .host("docs.api.com")
-                        .removePort(),
-                prettyPrint());
-    }
-
-    public static OperationResponsePreprocessor getDocumentResponse() {
-        return preprocessResponse(prettyPrint());
-    }
-
-}
