--- anonymous-board/src/test/java/com/example/anonymousboard/util/DocumentFormatGenerator.java
@@ -1,13 +1,0 @@
-package com.example.anonymousboard.util;
-
-
-import static org.springframework.restdocs.snippet.Attributes.key;
-
-import org.springframework.restdocs.snippet.Attributes;
-
-public class DocumentFormatGenerator {
-
-    public static Attributes.Attribute getConstraints(final String key, final String value) {
-        return key(key).value(value);
-    }
-}
